# apps/sandbox/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hyper Backend API
  description: |
    Backend API for the terminal-first, multiplayer agentic coding platform.

    This API manages:
    - **Sessions**: Isolated sandbox environments
    - **PTYs**: Pseudo-terminals within sessions
    - **Agents**: AI coding agents (Claude Code, Codex CLI)
    - **Filesystem**: Scoped file operations within /workspace

    ## Multiplayer Model
    - Sessions are single-tenant sandboxes
    - PTYs support multiple viewers but only one controller
    - Turn-taking is enforced via WebSocket control messages

    ## WebSocket Protocol
    PTY and Agent WebSocket connections use:
    - **Binary frames**: Raw terminal I/O
    - **Text frames (JSON)**: Control messages (resize, turn-taking)
  version: 1.0.0
  contact:
    name: Hyper AI
    url: https://hyper.ai
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://{sandbox}.fly.dev
    description: Fly.io sandbox
    variables:
      sandbox:
        default: hyper-sandbox
        description: Sandbox machine name
...
paths:
...
  /sessions/{sessionId}/agent:
    post:
      tags: [Agent]
      summary: Start the agent
      description: |
        Starts an AI coding agent in the session.
        Only one agent can run per session.
      operationId: startAgent
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '201':
          description: Agent started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Agent already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Agent]
      summary: Get agent status
      description: Returns the current agent state
      operationId: getAgent
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Agent status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '404':
          description: Session or agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      description: Session UUID
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: session not found
...
    AgentResponse:
      type: object
      required: [id, state]
      properties:
        id:
          type: string
          example: session-id-agent
        state:
          type: string
          enum: [running, paused, stopped]
          example: running
...